requirements.txt
streamlit>=1.27.0
google-cloud-storage>=2.14.0
google-auth>=2.28.0
portalocker
gcsfs
st-files-connection
pandas>=1.5.0
plotly>=5.0.0
